Move default annotated text implementations to default project
