+++ src/components/Text/Annotated/utils/isAnnotation.ts
@@ -1,7 +1,7 @@
 import { AnnotationSegment, NestedAnnotationSegment } from "../core";
 import { AnnoRepoBody } from "../../../../model/AnnoRepoAnnotation.ts";
 
-export function isProjectAnnotation(
+export function isAnnotation(
   toTest: AnnotationSegment,
 ): toTest is NestedAnnotationSegment<AnnoRepoBody> {
   return toTest.type === "annotation";
