+++ src/projects/default/annotation/defaultAnnotatedTextConfig.ts
@@ -1,12 +1,12 @@
 import { AnyAnnotatedTextConfig } from "../../../components/Text/Annotated/core";
-import { DefaultHighlightAnnotations } from "../../../components/Text/Annotated/common/highlight/DefaultHighlightAnnotations";
-import { DefaultMarkerAnnotation } from "../../../components/Text/Annotated/common/marker/DefaultMarkerAnnotation";
-import { DefaultSegmentGroup } from "../../../components/Text/Annotated/common/group/DefaultSegmentGroup";
-import { DefaultNestedAnnotation } from "../../../components/Text/Annotated/common/nested/DefaultNestedAnnotation.tsx";
+import { DefaultHighlights } from "./highlight/DefaultHighlights.tsx";
+import { DefaultMarker } from "./marker/DefaultMarker.tsx";
+import { DefaultGroup } from "./group/DefaultGroup.tsx";
+import { DefaultNested } from "./nested/DefaultNested.tsx";
 
 export const defaultAnnotatedTextConfig: AnyAnnotatedTextConfig = {
-  Nested: DefaultNestedAnnotation,
-  Highlight: DefaultHighlightAnnotations,
-  Marker: DefaultMarkerAnnotation,
-  Group: DefaultSegmentGroup,
+  Nested: DefaultNested,
+  Highlight: DefaultHighlights,
+  Marker: DefaultMarker,
+  Group: DefaultGroup,
 };
